ui/pyproject.toml
[tool.poetry]
name = "rfm-ui"
version = "0.1.0"
description = "Interactive UI for RFM Architecture Visualization"
authors = ["Pablo Sanchez <pablo.snchz@hotmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
dearpygui = "^1.10.1"
numpy = "^1.24.0"
pillow = "^10.0.0"
psutil = "^5.9.5"
pyperclip = "^1.8.2"
pyyaml = "^6.0.1"
numba = "^0.57.1"
matplotlib = "^3.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rfm-ui = "rfm_ui.main:main"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true